+++ src/main/java/com/task/taskreminder/repository/UserRepository.java
@@ -3,7 +3,8 @@ package com.task.taskreminder.repository;
 import com.task.taskreminder.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> 
+{
 
     User findByEmail(String email);
